+++ src/main/java/roomescape/dto/ReservationRequest.java
@@ -1,9 +1,10 @@
-package roomescape;
+package roomescape.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import java.time.LocalTime;
+import roomescape.entity.Reservation;
 
 public record ReservationRequest(
         @NotBlank String name,
