+++ src/main/java/roomescape/entity/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.entity;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
