+++ src/main/java/roomescape/repository/ReservationRepository.java
@@ -1,10 +1,11 @@
-package roomescape;
+package roomescape.repository;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.stereotype.Repository;
+import roomescape.entity.Reservation;
 
 @Repository
 public class ReservationRepository {
